requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0
coverage>=7.4